wip: update reducer slice for cart
